+++ engine/runtime/src/main/java/org/enso/interpreter/node/callable/dispatch/InvokeFunctionNode.java
@@ -76,7 +76,6 @@ public abstract class InvokeFunctionNode extends BaseNode {
     functionCallInstrumentationNode.execute(
         callerFrame,
         function,
-        callerInfo,
         mappedArguments.getState(),
         mappedArguments.getSortedArguments());
     return curryNode.execute(
